Revise/fix particle emitter position and direction information
The greatest issue was that positions were being stored in
"meters," whereas the rest of the active object system uses
the raw position of the object to Irrlicht (meter position * BS).
This is now fixed.

In addition, the direction/velocity of particles is no longer
multiplied by the size of a node (BS).  Finally, a typo in a comment
has been eliminated :) .